refactor: move to src folder
