+++ src/copy.js
